server/requirements.txt
uvicorn
fastapi
apiclient
pymongo[srv]
python-dotenv
python-multipart
pyjwt
